split experiment and visualisation, add config files for experiments
