+++ experiments/breadth_first/breadth_first_experiment.json
@@ -1,0 +1,20 @@
+{
+    "data_folder": "nl",
+    "scheme": 1,
+    "runs": [
+        {
+            "function": "breadth_first",
+            "params": {
+                "transmitter_count": 4,
+                "output_file": "results/breadth_first/breadth_first.json"
+            }
+        },
+        {
+            "function": "best_first",
+            "params": {
+                "transmitter_count": 4,
+                "output_file": "results/breadth_first/best_first.json"
+            }
+        }
+    ]
+}
