+++ experiments/depth_first/depth_first_experiment.json
@@ -1,0 +1,20 @@
+{
+    "data_folder": "nl",
+    "scheme": 1,
+    "runs": [
+        {
+            "function": "depth_first",
+            "params": {
+                "transmitter_count": 4,
+                "output_file": "results/depth_first/depth_first.json"
+            }
+        },
+        {
+            "function": "branchandbound",
+            "params": {
+                "transmitter_count": 4,
+                "output_file": "results/depth_first/branchandbound.json"
+            }
+        }
+    ]
+}
