+++ experiments/greedy/greedy_experiment.json
@@ -1,0 +1,19 @@
+{
+    "data_folder": "nl",
+    "scheme": 1,
+    "runs": [
+        {
+            "function": "random_greedy",
+            "params": {
+                "iterations": 10000,
+                "output_file": "results/greedy/random_greedy.csv"
+            }
+        },
+        {
+            "function": "greedy",
+            "params": {
+                "output_file": "results/greedy/greedy.csv"
+            }
+        }
+    ]
+}
