+++ experiments/greedy/greedy_experiment.py
@@ -1,57 +1,24 @@
 from radio_russia.algorithms import greedy as gr
-import matplotlib.pyplot as plt
 import csv
 
 
-def random_greedy(graph, transmitters):
+def random_greedy(graph, transmitters, scheme=1, iterations=10000,
+                   output_file="results/greedy/random_greedy.csv"):
     results = []
-    for _ in range(10000):
-        random_greedy_graph = gr.RandomGreedy(graph, transmitters.get_scheme((1)))
+    for _ in range(iterations):
+        random_greedy_graph = gr.RandomGreedy(graph, transmitters.get_scheme(scheme))
         random_greedy_graph.run()
         results.append((random_greedy_graph.graph.calculate_value(), random_greedy_graph.graph.to_json()))
 
-    with open("results/greedy/random_greedy.csv", 'w', newline='') as output_file:
+    with open(output_file, 'w', newline='') as output_file:
         result_writer = csv.writer(output_file, delimiter=',')
         for result in results:
             result_writer.writerow(result)
 
 
-def greedy(graph, transmitters):
-    greedy_graph = gr.Greedy(graph, transmitters.get_scheme((1)))
+def greedy(graph, transmitters, scheme=1, output_file="results/greedy/greedy.csv"):
+    greedy_graph = gr.Greedy(graph, transmitters.get_scheme(scheme))
     greedy_graph.run()
-    with open("results/greedy/greedy.csv", 'w', newline='') as output_file:
+    with open(output_file, 'w', newline='') as output_file:
         result_writer = csv.writer(output_file, delimiter=',')
         result_writer.writerow((greedy_graph.graph.calculate_value(), greedy_graph.graph.to_json()))
-
-
-def random_greedy_graph():
-    fig, ax = plt.subplots()
-    with open("results/greedy/random_greedy.csv", 'r') as input_file:
-        result_reader = csv.reader(input_file, delimiter=',')
-        results = [int(value) for value, _ in result_reader]
-
-    ax.hist(results)
-    ax.set_title('Random Greedy Solutions')
-    ax.set_xlabel('Total Costs')
-    ax.set_ylabel('Number of Solutions')
-    fig.savefig("results/greedy/random_greedy.png")
-
-
-def base_vs_random_graph():
-    fig, ax = plt.subplots()
-    
-    with open("results/greedy/random_greedy.csv", 'r') as input_file:
-        result_reader = csv.reader(input_file, delimiter=',')
-        results = [int(value) for value, _ in result_reader]
-
-    with open("results/greedy/greedy.csv", 'r') as input_file:
-        result_reader = csv.reader(input_file, delimiter=',')
-        greedy_result = [int(value) for value, _ in result_reader]
-
-    ax.hist(results, label='Random')
-    ax.vlines(greedy_result, 0, 4500, colors=['red'], label='Base')
-    ax.legend(loc='upper right')
-    ax.set_title('Base Greedy vs Random Greedy')
-    ax.set_xlabel('Total Costs')
-    ax.set_ylabel('Number of Solutions')
-    fig.savefig("results/greedy/base_vs_random.png")
