+++ experiments/hillclimber/hillclimber_experiment.json
@@ -1,0 +1,30 @@
+{
+    "data_folder": "nl",
+    "scheme": 1,
+    "runs": [
+        {
+            "function": "hillclimb",
+            "params": {
+                "iterations": 1000,
+                "output_file": "results/hillclimber/hillclimber.csv"
+            }
+        },
+        {
+            "function": "hillclimber_averages",
+            "params": {
+                "runs": 100,
+                "iterations": 1000,
+                "output_file": "results/hillclimber/hillclimber_averages.csv"
+            }
+        },
+        {
+            "function": "hillclimber_xopt_comparison",
+            "params": {
+                "runs": 100,
+                "iterations": 500,
+                "mutate_nodes_range": [1, 6],
+                "output_file_template": "results/hillclimber/hillclimber_xopt_{n}.csv"
+            }
+        }
+    ]
+}
