+++ experiments/simulatedannealing/simulatedannealing_experiment.json
@@ -1,0 +1,30 @@
+{
+    "data_folder": "nl",
+    "scheme": 1,
+    "runs": [
+        {
+            "function": "simulateannealing",
+            "params": {
+                "iterations": 100,
+                "output_file": "results/simulatedannealing/simulatedannealing.csv"
+            }
+        },
+        {
+            "function": "simulatedannealing_averages",
+            "params": {
+                "runs": 100,
+                "iterations": 500,
+                "output_file": "results/simulatedannealing/simulatedannealing_averages.csv"
+            }
+        },
+        {
+            "function": "simulatedannealing_temperature_comparisons",
+            "params": {
+                "runs": 100,
+                "iterations": 500,
+                "temperature_range": [1, 60, 10],
+                "output_file_template": "results/simulatedannealing/simulatedannealing_temp_{n}.csv"
+            }
+        }
+    ]
+}
