+++ main.py
@@ -14,6 +14,13 @@ from experiments.breadth_first import breadth_first_experiment
 from experiments.hillclimber import hillclimber_experiment
 from experiments.simulatedannealing import simulatedannealing_experiment
 
+from visualisation.random import random_visualisation
+from visualisation.greedy import greedy_visualisation
+from visualisation.depth_first import depth_first_visualisation
+from visualisation.breadth_first import breadth_first_visualisation
+from visualisation.hillclimber import hillclimber_visualisation
+from visualisation.simulatedannealing import simulatedannealing_visualisation
+
 from radio_russia.visualisation import visualise as vis
 
 if __name__ == "__main__":
@@ -34,7 +41,7 @@ if __name__ == "__main__":
     #       f"{random_graph.calculate_value()}")
 
     # random_experiment.baseline(test_graph, transmitters)
-    # random_experiment.baseline_graph()
+    # random_visualisation.baseline_graph()
 
 
     # --------------------------- Greedy ---------------------------------------
@@ -53,8 +60,8 @@ if __name__ == "__main__":
 
     # greedy_experiment.random_greedy(test_graph, transmitters)
     # greedy_experiment.greedy(test_graph, transmitters)
-    # greedy_experiment.random_greedy_graph()
-    # greedy_experiment.base_vs_random_graph()
+    # greedy_visualisation.random_greedy_graph()
+    # greedy_visualisation.base_vs_random_graph()
 
     # --------------------------- Depth First ----------------------------------
     # NOTE: We use [0:4] to only use the first four transmitters, which makes this
@@ -69,9 +76,9 @@ if __name__ == "__main__":
 
     # depth_first_experiment.depth_first(test_graph, transmitters)
     # depth_first_experiment.branchandbound(test_graph, transmitters)
-    # depth_first_experiment.depth_first_table()
-    # depth_first_experiment.depth_first_memory_graph()
-    # depth_first_experiment.depth_first_memory_comparison()
+    # depth_first_visualisation.depth_first_table()
+    # depth_first_visualisation.depth_first_memory_graph()
+    # depth_first_visualisation.depth_first_memory_comparison()
 
     # --------------------------- Breadth First --------------------------------
     # Note: this WILL crash on any of the maps provided, but should work for
@@ -85,11 +92,11 @@ if __name__ == "__main__":
 
     # breadth_first_experiment.breadth_first(test_graph, transmitters)
     # breadth_first_experiment.best_first(test_graph, transmitters)
-    # breadth_first_experiment.breadth_first_table()
-    # breadth_first_experiment.breadth_first_memory_graph()
-    # breadth_first_experiment.breadth_first_memory_comparison()
-    # breadth_first_experiment.constructive_memory_comparison()
-    # breadth_first_experiment.constructive_comparison()
+    # breadth_first_visualisation.breadth_first_table()
+    # breadth_first_visualisation.breadth_first_memory_graph()
+    # breadth_first_visualisation.breadth_first_memory_comparison()
+    # breadth_first_visualisation.constructive_memory_comparison()
+    # breadth_first_visualisation.constructive_comparison()
 
     # --------------------------- Hill Climber ---------------------------------
     # print("Setting up Hill Climber...")
@@ -103,14 +110,14 @@ if __name__ == "__main__":
 
     # hillclimber_experiment.hillclimb(test_graph, transmitters)
     # hillclimber_experiment.hillclimb_continue(test_graph, transmitters, "results/hillclimber/hillclimber.csv")
-    # hillclimber_experiment.hillclimb_graph()
+    # hillclimber_visualisation.hillclimb_graph()
 
     # hillclimber_experiment.hillclimber_averages(test_graph, transmitters)
-    # hillclimber_experiment.hillclimber_averages_graph()
-    # hillclimber_experiment.hillclimber_averages_filled_graph()
+    # hillclimber_visualisation.hillclimber_averages_graph()
+    # hillclimber_visualisation.hillclimber_averages_filled_graph()
 
     # hillclimber_experiment.hillclimber_xopt_comparison(test_graph, transmitters)
-    # hillclimber_experiment.hillclimber_xopt_comparison_graph()
+    # hillclimber_visualisation.hillclimber_xopt_comparison_graph()
 
     # --------------------------- Simulated Annealing --------------------------
     # It is very difficult to find a good starting temperature for SA. A rule to
@@ -135,11 +142,11 @@ if __name__ == "__main__":
 
     # simulatedannealing_experiment.simulateannealing(test_graph, transmitters)
     # simulatedannealing_experiment.simulateannealing_continue(test_graph, transmitters, "results/simulatedannealing/simulatedannealing.csv")
-    # simulatedannealing_experiment.simulateannealing_graph()
+    # simulatedannealing_visualisation.simulateannealing_graph()
 
     # simulatedannealing_experiment.simulatedannealing_averages(test_graph, transmitters)
-    # simulatedannealing_experiment.simulatedannealing_averages_graph()
-    # simulatedannealing_experiment.simulatedannealing_averages_filled_graph()
+    # simulatedannealing_visualisation.simulatedannealing_averages_graph()
+    # simulatedannealing_visualisation.simulatedannealing_averages_filled_graph()
 
     # simulatedannealing_experiment.simulatedannealing_temperature_comparisons(test_graph, transmitters)
-    simulatedannealing_experiment.simulatedannealing_temperature_comparisons_graph()
+    simulatedannealing_visualisation.simulatedannealing_temperature_comparisons_graph()
