+++ visualisation/greedy/greedy_visualisation.py
@@ -1,0 +1,35 @@
+import matplotlib.pyplot as plt
+import csv
+
+
+def random_greedy_graph():
+    fig, ax = plt.subplots()
+    with open("results/greedy/random_greedy.csv", 'r') as input_file:
+        result_reader = csv.reader(input_file, delimiter=',')
+        results = [int(value) for value, _ in result_reader]
+
+    ax.hist(results)
+    ax.set_title('Random Greedy Solutions')
+    ax.set_xlabel('Total Costs')
+    ax.set_ylabel('Number of Solutions')
+    fig.savefig("results/greedy/random_greedy.png")
+
+
+def base_vs_random_graph():
+    fig, ax = plt.subplots()
+
+    with open("results/greedy/random_greedy.csv", 'r') as input_file:
+        result_reader = csv.reader(input_file, delimiter=',')
+        results = [int(value) for value, _ in result_reader]
+
+    with open("results/greedy/greedy.csv", 'r') as input_file:
+        result_reader = csv.reader(input_file, delimiter=',')
+        greedy_result = [int(value) for value, _ in result_reader]
+
+    ax.hist(results, label='Random')
+    ax.vlines(greedy_result, 0, 4500, colors=['red'], label='Base')
+    ax.legend(loc='upper right')
+    ax.set_title('Base Greedy vs Random Greedy')
+    ax.set_xlabel('Total Costs')
+    ax.set_ylabel('Number of Solutions')
+    fig.savefig("results/greedy/base_vs_random.png")
